Rename translations to locales and rely on I18n to generate multiple locale files
