+++ lib/rails/generators/locales/templates/component.yml
